https://github.com/thehcginstitute-com/wp/issues/31

+++ wp-content/plugins/wp-rocket/inc/functions/files.php
@@ -22,7 +22,7 @@ function rocket_generate_advanced_cache_file( $advanced_cache = null ) {
 	 * @param bool True (default) to go ahead with advanced cache file generation; false to stop generation.
 	 */
 	if ( ! (bool) apply_filters( 'rocket_generate_advanced_cache_file', true ) ) {
-		return;
+		return false;
 	}
 
 	static $done = false;
@@ -32,7 +32,7 @@ function rocket_generate_advanced_cache_file( $advanced_cache = null ) {
 	}
 
 	if ( $done ) {
-		return;
+		return false;
 	}
 	$done = true;
 
@@ -41,7 +41,7 @@ function rocket_generate_advanced_cache_file( $advanced_cache = null ) {
 		$advanced_cache = $container->get( 'advanced_cache' );
 	}
 
-	rocket_put_content(
+	return rocket_put_content(
 		rocket_get_constant( 'WP_CONTENT_DIR' ) . '/advanced-cache.php',
 		$advanced_cache->get_advanced_cache_content()
 	);
@@ -58,7 +58,10 @@ function get_rocket_config_file() { // phpcs:ignore WordPress.NamingConventions.
 	$options = get_option( WP_ROCKET_SLUG );
 
 	if ( ! $options ) {
-		return;
+		return [
+			[],
+			'',
+		];
 	}
 
 	$buffer  = "<?php\n";
@@ -116,11 +119,11 @@ function get_rocket_config_file() { // phpcs:ignore WordPress.NamingConventions.
 		}
 
 		if ( 'secret_cache_key' === $option ) {
-			$buffer .= '$rocket_' . $option . ' = \'' . $value . "';\n";
+			$buffer .= '$rocket_' . $option . ' = \'' . sanitize_key( $value ) . "';\n";
 		}
 
 		if ( 'cache_reject_uri' === $option ) {
-			$buffer .= '$rocket_' . $option . ' = \'' . get_rocket_cache_reject_uri() . "';\n";
+			$buffer .= '$rocket_' . $option . ' = \'' . get_rocket_cache_reject_uri( true ) . "';\n";
 		}
 
 		if ( 'cache_query_strings' === $option ) {
@@ -151,6 +154,8 @@ function get_rocket_config_file() { // phpcs:ignore WordPress.NamingConventions.
 	$buffer .= '$rocket_cache_mandatory_cookies = ' . call_user_func( 'var_export', get_rocket_cache_mandatory_cookies(), true ) . ";\n";
 
 	$buffer .= '$rocket_cache_dynamic_cookies = ' . call_user_func( 'var_export', get_rocket_cache_dynamic_cookies(), true ) . ";\n";
+
+	$buffer .= '$rocket_permalink_structure = \'' . wp_slash( get_option( 'permalink_structure' ) ) . "';\n";
 
 	/** This filter is documented in inc/front/htaccess.php */
 	if ( apply_filters( 'rocket_url_no_dots', false ) ) {
@@ -228,6 +233,34 @@ function rocket_delete_config_file() {
 	list( $config_files_path ) = get_rocket_config_file();
 	foreach ( $config_files_path as $config_file ) {
 		rocket_direct_filesystem()->delete( $config_file );
+	}
+
+	// Bail out if WP Rocket is multisite.
+	if ( is_multisite() ) {
+		return;
+	}
+
+	try {
+		$config_dir = new FilesystemIterator( (string) rocket_get_constant( 'WP_ROCKET_CONFIG_PATH' ) );
+	} catch ( Exception $e ) {
+		return;
+	}
+
+	// Remove all files with php extension in the config folder.
+	foreach ( $config_dir as $file ) {
+		if ( ! $file->isFile() || 'php' !== strtolower( $file->getExtension() ) ) {
+			continue;
+		}
+
+		if ( 1 === substr_count( $file->getFilename(), '.' ) ) {
+			continue;
+		}
+
+		if ( false === strpos( rocket_direct_filesystem()->get_contents( $file->getPathname() ), '$rocket_cookie_hash' ) ) {
+			continue;
+		}
+
+		rocket_direct_filesystem()->delete( $file->getPathname() );
 	}
 }
 
@@ -331,6 +364,7 @@ function rocket_clean_minify( $extensions = [ 'js', 'css' ] ) {
 	$min_cache_path = rocket_get_constant( 'WP_ROCKET_MINIFY_CACHE_PATH' );
 	$min_path       = $min_cache_path . get_current_blog_id() . '/';
 	$iterator       = _rocket_get_cache_path_iterator( $min_path );
+
 	if ( false === $iterator ) {
 		return;
 	}
@@ -399,9 +433,23 @@ function rocket_clean_minify( $extensions = [ 'js', 'css' ] ) {
  * @return void
  */
 function rocket_clean_cache_busting( $extensions = [ 'js', 'css' ] ) {
-	$extensions = is_string( $extensions ) ? (array) $extensions : $extensions;
+	if ( empty( $extensions ) ) {
+		return;
+	}
 
-	$cache_busting_path = WP_ROCKET_CACHE_BUSTING_PATH . get_current_blog_id();
+	if ( is_string( $extensions ) ) {
+		$extensions = (array) $extensions;
+	}
+
+	$cache_busting_path = rocket_get_constant( 'WP_ROCKET_CACHE_BUSTING_PATH' ) . get_current_blog_id() . '/';
+	$iterator           = _rocket_get_cache_path_iterator( $cache_busting_path );
+
+	if ( false === $iterator ) {
+		return;
+	}
+
+	$filesystem         = rocket_direct_filesystem();
+	$busting_path_regex = str_replace( '/', '\/', $cache_busting_path );
 
 	if ( ! rocket_direct_filesystem()->is_dir( $cache_busting_path ) ) {
 		rocket_mkdir_p( $cache_busting_path );
@@ -417,20 +465,6 @@ function rocket_clean_cache_busting( $extensions = [ 'js', 'css' ] ) {
 		return;
 	}
 
-	try {
-		$dir = new RecursiveDirectoryIterator( $cache_busting_path, FilesystemIterator::SKIP_DOTS );
-	} catch ( UnexpectedValueException $e ) {
-		// No logging yet.
-		return;
-	}
-
-	try {
-		$iterator = new RecursiveIteratorIterator( $dir, RecursiveIteratorIterator::CHILD_FIRST );
-	} catch ( Exception $e ) {
-		// No logging yet.
-		return;
-	}
-
 	foreach ( $extensions as $ext ) {
 		/**
 		 * Fires before the cache busting files are deleted
@@ -442,13 +476,13 @@ function rocket_clean_cache_busting( $extensions = [ 'js', 'css' ] ) {
 		do_action( 'before_rocket_clean_busting', $ext ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
 
 		try {
-			$files = new RegexIterator( $iterator, '#.*\.' . $ext . '#', RegexIterator::GET_MATCH );
-			foreach ( $files as $file ) {
-				rocket_direct_filesystem()->delete( $file[0] );
-			}
-		} catch ( InvalidArgumentException $e ) {
-			// No logging yet.
+			$entries = new RegexIterator( $iterator, "/{$busting_path_regex}.*\.{$ext}/" );
+		} catch ( Exception $e ) {
 			return;
+		}
+
+		foreach ( $entries as $entry ) {
+			$filesystem->delete( $entry->getPathname() );
 		}
 
 		/**
@@ -461,22 +495,39 @@ function rocket_clean_cache_busting( $extensions = [ 'js', 'css' ] ) {
 		do_action( 'after_rocket_clean_cache_busting', $ext ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
 	}
 
-	try {
-		foreach ( $iterator as $item ) {
-			if ( rocket_direct_filesystem()->is_dir( $item ) ) {
-				rocket_direct_filesystem()->delete( $item );
-			}
+	foreach ( $iterator as $item ) {
+		if ( $filesystem->is_dir( $item ) ) {
+			$filesystem->delete( $item );
 		}
-	} catch ( UnexpectedValueException $e ) {
-		// Log the error.
-		Logger::debug(
-			'Cache Busting folder structure contains a directory we cannot recurse into.',
-			[
-				'Full error',
-				'UnexpectedValueException' => $e->getMessage(),
-			]
-		);
 	}
+}
+
+/**
+ * Returns the right path when the post is trashed.
+ *
+ * @param array $parsed_url current parsed url.
+ * @param int   $post_id      ID from the post.
+ *
+ * @return array
+ */
+function rocket_maybe_find_right_trash_url( array $parsed_url, int $post_id ) {
+
+	$post = get_post( $post_id );
+
+	if ( ! $post || 'trash' !== $post->post_status ) {
+		return $parsed_url;
+	}
+
+	$post->post_status = 'publish';
+
+	$permalink = get_permalink( $post );
+
+	if ( ! $permalink ) {
+		return $parsed_url;
+	}
+
+	$new_permalink = str_replace( '__trashed', '', $permalink );
+	return get_rocket_parse_url( $new_permalink );
 }
 
 /**
@@ -491,8 +542,9 @@ function rocket_clean_cache_busting( $extensions = [ 'js', 'css' ] ) {
  *
  * @param string|array              $urls       URLs of cache files to be deleted.
  * @param WP_Filesystem_Direct|null $filesystem Optional. Instance of filesystem handler.
+ * @param bool                      $run_actions Run actions.
  */
-function rocket_clean_files( $urls, $filesystem = null ) {
+function rocket_clean_files( $urls, $filesystem = null, $run_actions = true ) {
 	$urls = (array) $urls;
 	if ( empty( $urls ) ) {
 		return;
@@ -511,25 +563,27 @@ function rocket_clean_files( $urls, $filesystem = null ) {
 		$filesystem = rocket_direct_filesystem();
 	}
 
-	/**
-	 * Fires before all cache files are deleted.
-	 *
-	 * @since  3.2.2
-	 *
-	 * @param array $urls The URLs corresponding to the deleted cache files.
-	 */
-	do_action( 'before_rocket_clean_files', $urls ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+	if ( $run_actions ) {
+		/**
+		 * Fires before all cache files are deleted.
+		 *
+		 * @since  3.2.2
+		 *
+		 * @param array $urls The URLs corresponding to the deleted cache files.
+		 */
+		do_action( 'before_rocket_clean_files', $urls ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+	}
 
 	foreach ( $urls as $url ) {
-
-		/**
-		 * Fires before the cache file is deleted.
-		 *
-		 * @since 1.0
-		 *
-		 * @param string $url The URL that the cache file to be deleted.
-		 */
-		do_action( 'before_rocket_clean_file', $url ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+		if ( $run_actions ) {
+			/**
+			 * Fires before the cache file is deleted.
+			 *
+			 * @param string $url The URL that the cache file to be deleted.
+			 * @since 1.0
+			 */
+			do_action( 'before_rocket_clean_file', $url ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+		}
 
 		if ( $url_no_dots ) {
 			$url = str_replace( '.', '_', $url );
@@ -537,28 +591,52 @@ function rocket_clean_files( $urls, $filesystem = null ) {
 
 		$parsed_url = get_rocket_parse_url( $url );
 
-		foreach ( _rocket_get_cache_dirs( $parsed_url['host'], $cache_path ) as $dir ) {
-			$entry = $dir . $parsed_url['path'];
-			// Skip if the dir/file does not exist.
-			if ( ! $filesystem->exists( $entry ) ) {
-				continue;
-			}
+		if ( ! empty( $parsed_url['host'] ) ) {
+			foreach ( _rocket_get_cache_dirs( $parsed_url['host'], $cache_path ) as $dir ) {
+				// Decode url path.
+				$url_chunks = explode( '/', $parsed_url['path'] );
+				$matches    = preg_grep( '/%/', $url_chunks );
 
-			if ( $filesystem->is_dir( $entry ) ) {
-				rocket_rrmdir( $entry, [], $filesystem );
-			} else {
-				$filesystem->delete( $entry );
+				if ( ! empty( $matches ) ) {
+					$parsed_url['path'] = rawurldecode( $parsed_url['path'] );
+				}
+
+				// Encode Non-latin characters if found in url path.
+				if ( false !== preg_match_all( '/(?<non_latin>[^\x00-\x7F]+)/', $parsed_url['path'], $matches ) ) {
+					$cb_encode_non_latin = function ( $non_latin ) {
+						return strtolower( rawurlencode( $non_latin ) );
+					};
+
+					$parsed_url['path'] = str_replace( $matches['non_latin'], array_map( $cb_encode_non_latin, $matches['non_latin'] ), $parsed_url['path'] );
+				}
+
+				$entry = $dir . $parsed_url['path'];
+
+				// Skip if the dir/file does not exist.
+				if ( ! $filesystem->exists( $entry ) ) {
+					continue;
+				}
+
+				if ( $filesystem->is_dir( $entry ) ) {
+					rocket_rrmdir( $entry, [], $filesystem );
+				} else {
+					$filesystem->delete( $entry );
+				}
 			}
 		}
+		if ( $run_actions ) {
+			/**
+			 * Fires after the cache file is deleted.
+			 *
+			 * @param string $url The URL that the cache file was deleted.
+			 * @since 1.0
+			 */
+			do_action( 'after_rocket_clean_file', $url ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+		}
+	}
 
-		/**
-		 * Fires after the cache file is deleted.
-		 *
-		 * @since 1.0
-		 *
-		 * @param string $url The URL that the cache file was deleted.
-		 */
-		do_action( 'after_rocket_clean_file', $url ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
+	if ( ! $run_actions ) {
+		return;
 	}
 
 	/**
@@ -613,10 +691,13 @@ function rocket_clean_home( $lang = '' ) {
 	do_action( 'before_rocket_clean_home', $root, $lang ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
 
 	// Delete homepage.
-	$files = glob( $root . '/{index,index-*}.{html,html_gzip}', GLOB_BRACE | GLOB_NOSORT );
+	$files = glob( $root . '/*', GLOB_NOSORT );
+
 	if ( $files ) {
-		foreach ( $files as $file ) { // no array map to use @.
-			rocket_direct_filesystem()->delete( $file );
+		foreach ( $files as $file ) {
+			if ( preg_match( '#/index(?:-.+\.|\.)html(?:_gzip)?$#', $file ) ) {
+				rocket_direct_filesystem()->delete( $file );
+			}
 		}
 	}
 
@@ -628,8 +709,16 @@ function rocket_clean_home( $lang = '' ) {
 		}
 	}
 
+	$param_dirs = glob( $root . '/#*', GLOB_NOSORT );
+
+	if ( $param_dirs ) {
+		foreach ( $param_dirs as $dir ) {
+			rocket_rrmdir( $dir );
+		}
+	}
+
 	// Remove the hidden empty file for mobile detection on NGINX with the Rocket NGINX configuration.
-	$nginx_mobile_detect_files = glob( $root . '/.mobile-active', GLOB_BRACE | GLOB_NOSORT );
+	$nginx_mobile_detect_files = glob( $root . '/.mobile-active', GLOB_NOSORT );
 	if ( $nginx_mobile_detect_files ) {
 		foreach ( $nginx_mobile_detect_files as $nginx_mobile_detect_file ) { // no array map to use @.
 			rocket_direct_filesystem()->delete( $nginx_mobile_detect_file );
@@ -637,7 +726,7 @@ function rocket_clean_home( $lang = '' ) {
 	}
 
 	// Remove the hidden empty file for webp.
-	$nowebp_detect_files = glob( $root . '/.no-webp', GLOB_BRACE | GLOB_NOSORT );
+	$nowebp_detect_files = glob( $root . '/.no-webp', GLOB_NOSORT );
 	if ( $nowebp_detect_files ) {
 		foreach ( $nowebp_detect_files as $nowebp_detect_file ) { // no array map to use @.
 			rocket_direct_filesystem()->delete( $nowebp_detect_file );
@@ -663,6 +752,10 @@ function rocket_clean_home( $lang = '' ) {
  * @return void
  */
 function rocket_clean_home_feeds() {
+
+	if ( ! has_filter( 'rocket_cache_reject_uri', 'wp_rocket_cache_feed' ) ) {
+		return;
+	}
 
 	$urls   = [];
 	$urls[] = get_feed_link();
@@ -710,10 +803,17 @@ function rocket_clean_home_feeds() {
  * @param WP_Filesystem_Direct|null $filesystem Optional. Instance of filesystem handler.
  */
 function rocket_clean_domain( $lang = '', $filesystem = null ) {
+	if ( did_action( 'rocket_after_clean_domain' ) ) {
+		return;
+	}
+
+	if ( rocket_is_importing() ) {
+		return;
+	}
+
 	$urls = ( ! $lang || is_object( $lang ) || is_array( $lang ) || is_int( $lang ) )
 		? (array) get_rocket_i18n_uri()
 		: (array) get_rocket_i18n_home_url( $lang );
-
 	/**
 	 * Filter URLs to delete all caching files from a domain.
 	 *
@@ -725,7 +825,7 @@ function rocket_clean_domain( $lang = '', $filesystem = null ) {
 	$urls = (array) apply_filters( 'rocket_clean_domain_urls', $urls, $lang );
 	$urls = array_filter( $urls );
 	if ( empty( $urls ) ) {
-		return;
+		return false;
 	}
 
 	/** This filter is documented in inc/front/htaccess.php */
@@ -736,6 +836,7 @@ function rocket_clean_domain( $lang = '', $filesystem = null ) {
 	if ( empty( $filesystem ) ) {
 		$filesystem = rocket_direct_filesystem();
 	}
+
 	foreach ( $urls as $url ) {
 		$parsed_url = get_rocket_parse_url( $url );
 
@@ -780,14 +881,19 @@ function rocket_clean_domain( $lang = '', $filesystem = null ) {
 		 * @param string $url  The home url.
 		 */
 		do_action( 'after_rocket_clean_domain', $root, $lang, $url ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals
-		# 2023-12-28 Dmitrii Fediuk https://upwork.com/fl/mage2pro
-		# 1) "«Error: No cache files are cleared» on `wp rocket clean`":
-		# https://github.com/thehcginstitute-com/wp/issues/15
-		# 2) "How did I fix «Error: No cache files are cleared» on `wp rocket clean`?": https://df.tips/t/2093
-		# 3) https://github.com/wp-media/wp-rocket-cli/blob/v1.3/command.php#L283-L295
-		# 4) https://github.com/wp-media/wp-rocket/blob/v3.15.5/inc/functions/files.php#L892-L892
-		return true;
 	}
+
+	/**
+	 * Fires after all cache files was deleted.
+	 *
+	 * @since 3.15.5
+	 *
+	 * @param string $lang The current lang to purge.
+	 * @param array|string[] $urls  All urls to clean.
+	 */
+	do_action( 'rocket_after_clean_domain', $lang, $urls );
+
+	return true;
 }
 
 /**
@@ -884,7 +990,7 @@ function rocket_clean_user( $user_id, $lang = '' ) {
 		return;
 	}
 
-	$user_key = $user->user_login . '-' . get_rocket_option( 'secret_cache_key' );
+	$user_key = rawurlencode( $user->user_login ) . '-' . get_rocket_option( 'secret_cache_key' );
 
 	foreach ( $urls as $url ) {
 		$parse_url = get_rocket_parse_url( $url );
@@ -894,7 +1000,9 @@ function rocket_clean_user( $user_id, $lang = '' ) {
 			$parse_url['host'] = str_replace( '.', '_', $parse_url['host'] );
 		}
 
-		$root = rocket_get_constant( 'WP_ROCKET_CACHE_PATH' ) . $parse_url['host'] . '-' . $user_key . '*' . $parse_url['path'];
+		$cache_dir = $parse_url['host'] . '-' . strtolower( $user_key );
+		$cache_dir = $cache_dir . $parse_url['path'];
+		$root      = rocket_get_constant( 'WP_ROCKET_CACHE_PATH' ) . $cache_dir;
 
 		/**
 		 * Fires before all caching files are deleted for a specific user
@@ -1058,7 +1166,7 @@ function rocket_rrmdir( $dir, array $dirs_to_preserve = [], $filesystem = null )
  *
  * @since 2.10
  *
- * @return object WP_Filesystem_Direct instance
+ * @return WP_Filesystem_Direct WP_Filesystem_Direct instance
  */
 function rocket_direct_filesystem() {
 	require_once ABSPATH . 'wp-admin/includes/class-wp-filesystem-base.php';
@@ -1082,15 +1190,17 @@ function rocket_mkdir( $dir ) {
 /**
  * Recursive directory creation based on full path.
  *
+ * @param string                    $target path to the directory we want to create.
+ * @param WP_Filesystem_Direct|null $filesystem WordPress filesystem.
+ * @return bool True if directory is created/exists, false otherwise
  * @since 1.3.4
  *
  * @source wp_mkdir_p() in /wp-includes/functions.php
- *
- * @param string $target path to the directory we want to create.
- * @return bool True if directory is created/exists, false otherwise
  */
-function rocket_mkdir_p( $target ) {
+function rocket_mkdir_p( $target, $filesystem = null ) {
 	$wrapper = null;
+
+	$filesystem = $filesystem ?: rocket_direct_filesystem();
 
 	if ( rocket_is_stream( $target ) ) {
 		list( $wrapper, $target ) = explode( '://', $target, 2 );
@@ -1110,20 +1220,20 @@ function rocket_mkdir_p( $target ) {
 		$target = '/';
 	}
 
-	if ( rocket_direct_filesystem()->exists( $target ) ) {
-		return rocket_direct_filesystem()->is_dir( $target );
+	if ( $filesystem->exists( $target ) ) {
+		return $filesystem->is_dir( $target );
 	}
 
 	// Attempting to create the directory may clutter up our display.
 	if ( rocket_mkdir( $target ) ) {
 		return true;
-	} elseif ( rocket_direct_filesystem()->is_dir( dirname( $target ) ) ) {
+	} elseif ( $filesystem->is_dir( dirname( $target ) ) ) {
 		return false;
 	}
 
 	// If the above failed, attempt to create the parent node, then try again.
-	if ( ( '/' !== $target ) && ( rocket_mkdir_p( dirname( $target ) ) ) ) {
-		return rocket_mkdir_p( $target );
+	if ( ( '/' !== $target ) && ( rocket_mkdir_p( dirname( $target ), $filesystem ) ) ) {
+		return rocket_mkdir_p( $target, $filesystem );
 	}
 
 	return false;
@@ -1255,7 +1365,6 @@ function _rocket_get_dir_files_by_regex( $dir, $regex ) { // phpcs:ignore WordPr
 	} catch ( Exception $e ) {
 		return [];
 	}
-
 }
 
 /**
